Move tests.utils -> utils.
