+++ ivmodels/tests/pulse.py
@@ -2,8 +2,7 @@ import numpy as np
 import scipy
 
 from ivmodels.quadric import Quadric
-from ivmodels.tests.utils import _check_test_inputs
-from ivmodels.utils import oproj, proj
+from ivmodels.utils import _check_test_inputs, oproj, proj
 
 
 def pulse_test(Z, X, y, beta, C=None, W=None, D=None, fit_intercept=True):
